refactor(tracing): change dir
